Kaczmarczyk&Jędraczka: few test funciton added, correction of name corrrelation -> correlation, pdf files with plots for mean energy correlation, mean cluster size and its connection to correlation length

+++ trunk/IsingModelSimulations5_2/Tools.cpp
@@ -12,7 +12,7 @@
 
 using namespace std;
 
-void Display(vector<int> S) {
+void Display(vector<short> S) {
     for (int i = 0; i < S.size(); i++) {
         (S[i] == 1) ? cout << " o" : cout << " x";
     }
